Implement sys$exit and proper process context switching.
This implement the sys$exit syscall which conforms to the System V i386 ABI for running processes.
Before this, processes were basically considered to be functions but this is unsafe and doesn't conform to the ABI. 
There's now proper context switching between the kernel and processes and processes get their own ("userspace") stack.